requirements/requirements.txt
# This file was autogenerated by uv via the following command:
#    uv pip compile --universal requirements/requirements.in --output-file requirements/requirements.txt
colorama==0.4.6 ; platform_system == 'Windows'
    # via tqdm
contourpy==1.3.3
    # via matplotlib
cycler==0.12.1
    # via matplotlib
fonttools==4.59.2
    # via matplotlib
kiwisolver==1.4.9
    # via matplotlib
matplotlib==3.10.6
    # via
    #   -r requirements/requirements.in
    #   seaborn
numpy==2.3.2
    # via
    #   -r requirements/requirements.in
    #   contourpy
    #   matplotlib
    #   pandas
    #   seaborn
packaging==25.0
    # via matplotlib
pandas==2.3.2
    # via seaborn
pillow==11.3.0
    # via
    #   -r requirements/requirements.in
    #   matplotlib
polars==1.32.3
    # via -r requirements/requirements.in
pyparsing==3.2.3
    # via matplotlib
python-dateutil==2.9.0.post0
    # via
    #   matplotlib
    #   pandas
pytz==2025.2
    # via pandas
seaborn==0.13.2
    # via -r requirements/requirements.in
six==1.17.0
    # via python-dateutil
tqdm==4.67.1
    # via -r requirements/requirements.in
tzdata==2025.2
    # via pandas
